Order code, structure files and add formatter
