+++ game/.prettierrc
@@ -1,0 +1,14 @@
+{
+    "tabWidth": 4,
+    "trailingComma": "all",
+    "printWidth": 80,
+    "semi": true,
+    "overrides": [
+        {
+            "files": "*.json",
+            "options": {
+                "tabWidth": 2
+            }
+        }
+    ]
+}
