+++ game/README.md
@@ -3,7 +3,6 @@
 - `src` - source code for your kaplay project
 - `dist` - distribution folder, contains your index.html, built js bundle and static assets
 
-
 ## Development
 
 ```sh
