+++ game/index.html
@@ -1,10 +1,11 @@
-<!DOCTYPE html>
+<!doctype html>
 <html>
-<head>
-<title>mygame</title>
-<link rel="stylesheet" href="style.css">
-</head>
-<body style="overflow:hidden">
-<script src="src/main.js" type="module"></script>
-</body>
+    <head>
+        <title>mygame</title>
+        <link rel="stylesheet" href="/style.css" />
+    </head>
+
+    <body style="overflow: hidden">
+        <script src="src/main.js" type="module"></script>
+    </body>
 </html>
