+++ game/public/style.css
@@ -14,8 +14,11 @@ canvas {
 body {
     margin: 0;
     background: var(--bg);
-    background-image:
-        linear-gradient(90deg, rgba(35, 9, 30, 0.8) 2px, transparent 1px),
+    background-image: linear-gradient(
+            90deg,
+            rgba(35, 9, 30, 0.8) 2px,
+            transparent 1px
+        ),
         linear-gradient(0deg, rgba(9, 35, 50, 0.8) 2px, transparent 1px);
     background-size: 30px 30px;
     display: flex;
