+++ game/src/assets.js
@@ -1,0 +1,8 @@
+// @ts-check
+
+import { k } from "./kaplay.js";
+
+k.loadFont("monogram", "/fonts/monogram.ttf", {
+    outline: 4,
+    filter: "linear",
+});
