+++ game/src/constants.js
@@ -1,0 +1,12 @@
+// @ts-check
+
+import dialogs from "./data/dialog.json";
+
+export const goalBlocks = 2;
+export const maxtime = 500;
+export const maxMistakes = 2;
+export const lineHeight = 24;
+export const charSpacing = 12;
+export const startmoveline = 5;
+export const marginvisiblebox = -40;
+export const jsonData = dialogs;
