+++ game/src/data/colorTags.json
@@ -1,0 +1,33 @@
+{
+  "0": {
+    "/C1": [
+      0, 7, 25, 27, 39, 42, 51, 54, 56, 59, 71, 83, 90, 92, 117, 123, 156, 162
+    ],
+    "/C2": [9, 16, 43, 45, 102, 104, 141, 141, 143],
+    "/C3": [
+      17, 24, 28, 36, 46, 50, 55, 55, 60, 61, 65, 68, 86, 89, 93, 101, 105, 116,
+      129, 140, 144, 155, 167, 170
+    ],
+    "/C4": [37, 38, 62, 63, 69, 70, 84, 85],
+    "/C5": [64, 64, 124, 128, 163, 166]
+  },
+  "1": {
+    "/C1": [
+      0, 8, 32, 37, 58, 61, 63, 66, 116, 119, 135, 138, 162, 165, 167, 170, 199,
+      201, 223, 225, 436, 441
+    ],
+    "/C2": [9, 20, 91, 95],
+    "/C3": [
+      21, 31, 38, 39, 42, 57, 62, 67, 72, 74, 76, 83, 86, 90, 96, 100, 102, 109,
+      112, 115, 120, 125, 128, 134, 139, 143, 146, 154, 161, 166, 171, 172, 175,
+      183, 186, 192, 195, 196, 202, 209, 211, 216, 221, 227, 229, 233, 268, 271,
+      281, 285, 311, 315, 328, 331, 348, 352, 377, 381, 387, 391, 422, 426, 433,
+      443, 452
+    ],
+    "/C4": [
+      40, 41, 73, 84, 110, 111, 126, 144, 145, 148, 149, 152, 173, 174, 184,
+      185, 193, 194, 210, 214, 231, 269, 283, 313, 330, 350, 379, 389, 424
+    ],
+    "/C5": [75, 101, 150, 151]
+  }
+}
