+++ game/src/kaplay.js
@@ -1,0 +1,8 @@
+// @ts-check
+
+import kaplay from "kaplay";
+
+export const k = kaplay({
+    background: [10, 10, 27, 0],
+    font: "monogram",
+});
