+++ game/src/scenes/endGame.js
@@ -1,0 +1,7 @@
+scene("endgame", () => {
+    const endgameLabel = add([
+        text("Analytics", { size: 48 }),
+        pos(center().x, center().y - 100),
+        anchor("center"),
+    ]);
+});
